+++ gbs.h
@@ -1,4 +1,4 @@
-/* $Id: gbs.h,v 1.4 2003/11/29 19:03:15 ranma Exp $
+/* $Id: gbs.h,v 1.5 2003/11/30 14:35:59 ranma Exp $
  *
  * gbsplay is a Gameboy sound player
  *
@@ -21,29 +21,29 @@ struct gbs_subsong_info {
 
 struct gbs {
 	uint8_t *buf;
-	uint32_t version;
-	uint32_t songs;
-	uint32_t defaultsong;
-	uint32_t load;
-	uint32_t init;
-	uint32_t play;
-	uint32_t stack;
-	uint32_t tma;
-	uint32_t tmc;
+	int version;
+	int songs;
+	int defaultsong;
+	uint16_t load;
+	uint16_t init;
+	uint16_t play;
+	uint16_t stack;
+	uint8_t tma;
+	uint8_t tmc;
 	char *title;
 	char *author;
 	char *copyright;
-	uint32_t codelen;
+	int codelen;
 	uint8_t *code;
 	uint8_t *exthdr;
-	uint32_t filesize;
+	size_t filesize;
 	uint32_t crc;
 	uint32_t crcnow;
 	struct gbs_subsong_info *subsong_info;
 	char *strings;
 	char v1strings[33*3];
 	uint8_t *rom;
-	uint32_t romsize;
+	int romsize;
 };
 
 struct gbs *gbs_open(char *name);
